- Added sentiment app icon - Moved from thorac to sentiment
