+++ app/src/main/AndroidManifest.xml
@@ -1,19 +1,19 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
-    package="com.example.thorac">
+    package="com.example.sentiment">
 
     <application
         android:allowBackup="true"
-        android:icon="@mipmap/ic_launcher"
+        android:icon="@mipmap/sentimentlogo"
         android:label="@string/app_name"
-        android:roundIcon="@mipmap/ic_launcher_round"
+        android:roundIcon="@mipmap/sentimentlogo_round"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-        <activity android:name=".ThreadMessages"></activity>
-        <activity android:name=".ThreadLists" />
-        <activity android:name=".NewThread" />
-        <activity android:name=".EmailAuth">
+        <activity android:name="com.example.sentiment.ThreadMessages"></activity>
+        <activity android:name="com.example.sentiment.ThreadLists" />
+        <activity android:name="com.example.sentiment.NewThread" />
+        <activity android:name="com.example.sentiment.EmailAuth">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
