+++ app/src/main/java/com/example/sentiment/ThreadListsAdapter.java
@@ -1,11 +1,9 @@
-package com.example.thorac;
+package com.example.sentiment;
 
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.TextView;
-
-import java.util.List;
 
 import androidx.annotation.NonNull;
 import androidx.recyclerview.widget.RecyclerView;
