+++ app/src/main/java/com/example/sentiment/ThreadMessagesAdapter.java
@@ -1,4 +1,4 @@
-package com.example.thorac;
+package com.example.sentiment;
 
 import android.view.LayoutInflater;
 import android.view.View;
@@ -6,7 +6,6 @@ import android.view.ViewGroup;
 import android.widget.TextView;
 
 import java.util.List;
-import java.util.stream.Collectors;
 
 import androidx.annotation.NonNull;
 import androidx.recyclerview.widget.RecyclerView;
