+++ app/src/main/java/com/example/sentiment/User.java
@@ -1,4 +1,4 @@
-package com.example.thorac;
+package com.example.sentiment;
 
 class User {
