+++ app/src/main/java/com/example/sentiment/UserMessage.java
@@ -1,4 +1,4 @@
-package com.example.thorac;
+package com.example.sentiment;
 
 class UserMessage {
