+++ app/src/main/res/mipmap-anydpi-v26/sentimentlogo.xml
@@ -1,0 +1,5 @@
+<?xml version="1.0" encoding="utf-8"?>
+<adaptive-icon xmlns:android="http://schemas.android.com/apk/res/android">
+    <background android:drawable="@drawable/sentiment_launcher_background"/>
+    <foreground android:drawable="@mipmap/sentiment_launcher_foreground"/>
+</adaptive-icon>
