+++ app/src/main/res/values/strings.xml
@@ -1,6 +1,6 @@
 <resources>
-    <string name="app_name">Thorac</string>
-    <string name="thorac">SENTiment</string>
+    <string name="app_name">SENTiment</string>
+    <string name="sentiment">SENTiment</string>
     <string name="sign_in">Sign In</string>
     <string name="register">Register</string>
     <string name="textview">TextView</string>
