+++ app/src/test/java/com/example/sentiment/ExampleUnitTest.java
@@ -1,4 +1,4 @@
-package com.example.thorac;
+package com.example.sentiment;
 
 import org.junit.Test;
